extract Lab trace fetch refs
