+++ src/core/runner/mod.rs
@@ -29,6 +29,7 @@ mod lab_selection;
 mod lab_workspaces;
 mod offload_changed_since;
 mod offload_metadata;
+mod origin_refs;
 mod resource_metrics;
 mod rig_materialization;
 mod session;
